requirements.txt
flask==2.0.2
python-dotenv==0.19.2
flask-restful==0.3.9
flask-marshmallow==0.14.0